requirements.txt
# Shared Requirements
mysql-connector-python